+++ Plugins/DreamShader/Source/DreamShaderEditor/Private/MaterialAssetGeneration/DreamShaderMaterialGeneratorPrivate.h
@@ -2,6 +2,7 @@
 
 #include "CoreMinimal.h"
 #include "DreamShaderTypes.h"
+#include "MaterialAssetGeneration/DreamShaderGeneratedAssetMetadata.h"
 
 #include "Materials/MaterialExpressionCustom.h"
 #include "Materials/MaterialExpressionObjectPositionWS.h"
@@ -310,11 +311,6 @@ namespace UE::DreamShader::Editor::Private
 		ECustomMaterialOutputType& OutReturnType,
 		bool& bOutReturnIsSubstrateMaterial,
 		FString& OutError);
-	FString BuildSourceHash(const FString& SourceText);
-	bool IsGeneratedAssetSourceCurrent(UObject* Asset, const FString& SourceFilePath, const FString& SourceHash);
-	void ApplySourceMetadata(UObject* Asset, const FString& SourceFilePath);
-	void ApplySourceMetadata(UObject* Asset, const FString& SourceFilePath, const FString& SourceHash);
-	bool SaveAssetPackage(UObject* Asset, FString& OutError);
 	UClass* ResolveMaterialExpressionClass(const FString& ClassSpecifier);
 	FProperty* FindMaterialExpressionArgumentProperty(UClass* ExpressionClass, const FString& ArgumentName);
 	bool IsMaterialExpressionInputProperty(const FProperty* Property);